requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.0.0

